Introduction of proxied models for secure hashing
We introduce proxies in order provide a layer of security within the
model. This is necessary because we want to re-hash the key that the
model objects use based on the user's secret and a public-like key which
is sent back and forth between the client and server.

This way a malicious user will not be able to assign other user's items
to themselves potentially discovering and sending the other user's item
key in place of their own.

With the double-hashing based on the user's secret, we never expose the
real key storing the data, so a malicious user would have to brute force
the entire hash space to get other another user's data.

Each of the models have a corresponding proxy to handle the re-hashing
of the key with the user's secret. For instance, the category and item
models have such proxies. Also, the relations of a category to a user
and an item to a category have proxies.

One note is that the show route was necessary in order to allow the
controller to use the item_url auto-generated method. We need to
determine if there is a better way to do this, or maybe we need the show
route, which would make this work.