Development of Oregano
The main library for handling the web-requests for indexing the URLs for
a given Oregano user, as well as the create functionality for view
relations and pinch relations has been added. In addition, the
functionality for the creation of an add relation of the URLs also has
been implemented.

Neo4j was chosen as the database for the system since it could
appropriate represent the relations between an Oregano user and the URLs
that user created. Each view and pinch is also seen as a relation in the
system and stored as such in the database.

Note that tests have been added for the functionality of the library
back-end code. The front-end JavaScript does not include tests. Also, a
manifest is used for the website for aggressive client-side caching of
assets.