Resolves issue #1 of unnecessary DB queries
When the URL index was fetched from the server, there were two queries
being made for each URL node. This lead to far too many queries than
necessary.

The resolution is to formulate one query to fetch the extra information
regarding the relation types when the URL nodes are fetched. Therefore,
issue #1 is resolved since only the needed query is now made on the DB.