Resolves issue #2 by adding the URL's title
The server will send an HTTP request fetching the URL's webpage and
parse the HTML accordingly to extract the title. This title is stored in
the database and used when the URLs are rendered.

Note that in the JQDeferred module, the isResolved and isRejected
methods were deprecated, so now the state method is used to determine
the whether the promise is resolved or rejected in the test cases.